Moving SpannerTransactionDefinition to core driver module
